feat hook for a reusable component of the api call and added images to the json file
