+++ public/events.json
@@ -7,12 +7,12 @@
     "hora": "19:00",
     "descripcion": "Concierto de rock en vivo",
     "precio": 10000,
-    "imagen_evento": "https://dummyimage.com/420x320/ff7f7f/333333.png&text=Sample",
+    "imagen_evento": "https://images.pexels.com/photos/167636/pexels-photo-167636.jpeg?auto=compress&cs=tinysrgb&w=400",
     "region": "Región Metropolitana",
     "comuna": "Las Condes",
     "direccion": "Av. Apoquindo 6000",
     "comentarios": "¡Excelente concierto!",
-    "likes": 5
+    "likes": 5400
   },
   {
     "id": 2,
@@ -22,7 +22,7 @@
     "hora": "12:00",
     "descripcion": "Feria de comida con variedad de platos",
     "precio": 5000,
-    "imagen_evento": "https://dummyimage.com/420x320/ff7f7f/333333.png&text=Sample",
+    "imagen_evento": "https://images.pexels.com/photos/15211938/pexels-photo-15211938.jpeg?auto=compress&cs=tinysrgb&w=400",
     "region": "Región Metropolitana",
     "comuna": "Providencia",
     "direccion": "Av. Nueva Providencia 2155",
@@ -37,7 +37,7 @@
     "hora": "20:00",
     "descripcion": "Noche de comedia con los mejores comediantes",
     "precio": 8000,
-    "imagen_evento": "https://dummyimage.com/420x320/ff7f7f/333333.png&text=Sample",
+    "imagen_evento": "https://images.pexels.com/photos/713149/pexels-photo-713149.jpeg?auto=compress&cs=tinysrgb&w=400",
     "region": "Región Metropolitana",
     "comuna": "La Florida",
     "direccion": "Av. Vicuña Mackenna 6100",
@@ -52,7 +52,7 @@
     "hora": "14:00",
     "descripcion": "Fiesta en la piscina con música en vivo",
     "precio": 15000,
-    "imagen_evento": "https://dummyimage.com/420x320/ff7f7f/333333.png&text=Sample",
+    "imagen_evento": "https://images.pexels.com/photos/587738/pexels-photo-587738.jpeg?auto=compress&cs=tinysrgb&w=400",
     "region": "Región Metropolitana",
     "comuna": "Ñuñoa",
     "direccion": "Av. Irarrázaval 4949",
@@ -61,17 +61,17 @@
   },
   {
     "id": 5,
-    "tipo_evento": "fiesta",
-    "nombre_evento": "Fiesta en la playa",
+    "tipo_evento": "deportivo",
+    "nombre_evento": "30K de Santiago ",
     "fecha": "2023-03-25",
     "hora": "21:00",
-    "descripcion": "Fiesta de música electrónica en la playa",
+    "descripcion": "Carrera en bicicleta de 30k por la ciudad",
     "precio": 8000,
-    "imagen_evento": "https://dummyimage.com/420x320/ff7f7f/333333.png&text=Sample",
+    "imagen_evento": "https://images.pexels.com/photos/3771806/pexels-photo-3771806.jpeg?auto=compress&cs=tinysrgb&w=400",
     "region": "Región Metropolitana",
     "comuna": "Pudahuel",
     "direccion": "Av. San Pablo 7200",
-    "comentarios": "Muy buena música",
+    "comentarios": "Muy buen evento",
     "likes": 40
   },
   {
@@ -82,7 +82,7 @@
     "hora": "15:00",
     "descripcion": "Obra de teatro infantil sobre el espacio",
     "precio": 4000,
-    "imagen_evento": "https://dummyimage.com/420x320/ff7f7f/333333.png&text=Sample",
+    "imagen_evento": "https://images.pexels.com/photos/5801305/pexels-photo-5801305.jpeg?auto=compress&cs=tinysrgb&w=400",
     "region": "Región Metropolitana",
     "comuna": "La Florida",
     "direccion": "Av. Vicuña Mackenna 6100",
@@ -97,7 +97,7 @@
     "hora": "19:00",
     "descripcion": "Charla sobre el impacto del cambio climático en nuestra vida diaria",
     "precio": 5000,
-    "imagen_evento": "https://dummyimage.com/420x320/ff7f7f/333333.png&text=Sample",
+    "imagen_evento": "https://images.pexels.com/photos/2990647/pexels-photo-2990647.jpeg?auto=compress&cs=tinysrgb&w=400",
     "region": "Región Metropolitana",
     "comuna": "Santiago",
     "direccion": "Av. Libertador Bernardo O'Higgins 227",
@@ -112,7 +112,7 @@
     "hora": "09:00",
     "descripcion": "Carrera de 10k por la ciudad",
     "precio": 12000,
-    "imagen_evento": "https://dummyimage.com/420x320/ff7f7f/333333.png&text=Sample",
+    "imagen_evento": "https://images.pexels.com/photos/15723095/pexels-photo-15723095.jpeg?auto=compress&cs=tinysrgb&w=400",
     "region": "Región Metropolitana",
     "comuna": "Las Condes",
     "direccion": "Av. Apoquindo 4501",
@@ -127,7 +127,7 @@
     "hora": "16:00",
     "descripcion": "Exposición de arte moderno",
     "precio": 3000,
-    "imagen_evento": "https://dummyimage.com/420x320/ff7f7f/333333.png&text=Sample",
+    "imagen_evento": "https://images.pexels.com/photos/1839919/pexels-photo-1839919.jpeg?auto=compress&cs=tinysrgb&w=400",
     "region": "Región Metropolitana",
     "comuna": "Providencia",
     "direccion": "Av. Providencia 1208",
@@ -142,7 +142,7 @@
     "hora": "10:00",
     "descripcion": "Feria de productos locales y artesanías",
     "precio": 0,
-    "imagen_evento": "https://dummyimage.com/420x320/ff7f7f/333333.png&text=Sample",
+    "imagen_evento": "https://images.pexels.com/photos/4072329/pexels-photo-4072329.jpeg?auto=compress&cs=tinysrgb&w=400",
     "region": "Región Metropolitana",
     "comuna": "Ñuñoa",
     "direccion": "Av. Irarrázaval 2401",
